Changed project structure
