+++ runtime/CMakeLists.txt
@@ -4,6 +4,9 @@ project ("KTL Runtime Library")
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
+set(KTL_RUNTIME_INCLUDE_DIR "${KTL_RUNTIME_DIR}/include")
+set(KTL_RUNTIME_SOURCE_DIR "${KTL_RUNTIME_DIR}/src")
+
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}/lib")
 set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}/lib")
