--- runtime/src/exc_engine/stack_walker.h
@@ -1,6 +1,0 @@
-#pragma once
-#include <exc_typedefs.h>
-
-namespace ktl::crt::exc_engine {
-void stack_walk(record_t& exc_rec);
-}
