--- runtime/src/exc_engine/x64/invoke_funclet.h
@@ -1,4 +1,0 @@
-#pragma once
-#include <crt_attributes.h>
-
-EXTERN_C void* exc_engine_invoke_funclet(void* funclet, void* stack_frame);
