+++ runtime/src/exception_base.cpp
@@ -1,11 +1,7 @@
 #include <crt_attributes.h>
-#include <exc_engine_interface.h>
 #include <exception_base.h>
 #include <heap.h>
 
-using ktl::crt::exc_engine::cpu_context_t;
-using ktl::crt::exc_engine::disposition_t;
-using ktl::crt::exc_engine::record_t;
 
 namespace ktl::crt {
 using exception_index = uint16_t;
@@ -100,5 +96,3 @@ void cleanup_exception_environment() {
         exc_rec, frame_ptr, cpu_context, dispatcher_context);          \
   }
 
-DEFINE_CXX_FRAME_HANDLER(3)
-DEFINE_CXX_FRAME_HANDLER(4)
